build: change file naming, copy flow files
